Fix Docker build error from db config
Once I add the prometheus app, the collectstatic process becomes
upset with the db not being configured properly. This commit forces
the engine to be the 'dummy' one which is just fine for running the
collectstatic/compress process.

Also remove the app template finder as it generates errors during
the collectstatic process, and we're not using per-app static files
anyway

Change-Id: I62a0ee020832858244bc610825b7f8d0d591db2d