Improve multiple desktop (policy) check
The existing approach of checking the "latest" VMStatus is not
robust.  Sometimes we can get a situation where there is a live
Instance associated with an VMStatus that is not the latest one.

So now we will look at the Instance records themselves.  And also
do the check when unshelving.

Change-Id: I9a5a9f835f065c52db27311d3eb8daa05109b05e