Add support for OpenStack hypervisor-specific VNC console access
This support adds a new way to access the grahical console of any VDI
instance and provide it to web clients via Guacamole. In addition to the
available standard console access of the integrated RDP server from each
VDI instance, the hypervisor's VNC console from OpenStack can now be
used as an alternative. The alternative offers the advantage that the
graphic output can already be accessed during the boot process of a VDI
instance. In addition, this option saves resources, since no extra RDP
server has to run in each deployed VDI instance.

The new OpenStack VNC console access can be selected via the new
configuration option

  - OS_CONSOLE_SERVER = [ openstack_hypervisor | instance_builtin ]

by assigning the value 'openstack_hypervisor'. The prerequisite for
activating the OpenStack VNC console access is that an OpenStack-based
cloud is set up with the console type 'noVNC'.