Fix the logout button
In the OIDC case, I was redirecting the "/logout/" URL to
"/oidc/logout".  But redirects don't work for a POSTFix the logout
button

In the OIDC case, I was redirecting the "/logout/" URL to
"/oidc/logout".  But redirects don't work for a POST.

There is still the problem that Keycloak doesn't know that the user
has logged out, and retains the authentication.  But arguably that
is correct.

Change-Id: Iac236582190cd910c49471b076d467f4cfa80e7e