Desktop-type specific timeouts
Some desktop types take significantly longer to get to the
phone-home point.  Rather than setting a long launch time for
everything, this allows for desktop-type specific timeouts.

Also tweak the progress bar numbers so that the progress bar is
less likely to get "stuck at 100%" when booting Windows.  (There
is a Jira ticket to do this better.)

Change-Id: I1af49734817246b81d6d951d60ec08edfcd07fee