Tweaks to make admin pages load faster
The problem is that if a foreign key field is updatable, an object's
admin page will have a pulldown populated with all possible values
for the target object.  This gets BIG and slows down page loading.

Since it makes little / no sense to update these fields, we can
just label them all read-only in the admin interface

Change-Id: I814aaeb5c6ac04c89982de5fb9619453ba6a546f